requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
watchdog==2.1.6
Werkzeug==2.0.3
